+++ packages/extension-base/src/koni/background/handlers/Tabs.ts
@@ -177,8 +177,8 @@ export default class KoniTabs {
     }
   }
 
-  private metadataProvide (url: string, request: MetadataDef): Promise<boolean> {
-    return this.#koniState.injectMetadata(url, request);
+  private metadataProvide (url: string, request: MetadataDef): boolean {
+    return this.#koniState.injectMetadata(request);
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -591,14 +591,14 @@ export default class KoniTabs {
           chainId: chainId,
           rpcUrls: onlineData.rpc.filter((url) => (url.startsWith('https://'))),
           chainName: onlineData.name,
-          blockExplorerUrls: onlineData.explorers.map((explorer) => explorer.url),
+          blockExplorerUrls: onlineData.explorers?.map((explorer) => explorer.url),
           nativeCurrency: onlineData.nativeCurrency,
           requestId: id
         };
 
         await this.addEvmChain(id, url, { method: 'wallet_addEthereumChain', params: [chainData] });
       } else {
-        throw new EvmProviderError(EvmProviderErrorType.INVALID_PARAMS, 'This network is currently not supported');
+        throw new EvmProviderError(EvmProviderErrorType.NETWORK_NOT_SUPPORTED, 'This network is currently not supported');
       }
     }