+++ packages/extension-base/src/services/balance-service/helpers/subscribe/substrate/index.ts
@@ -17,7 +17,8 @@ import { getDefaultWeightV2 } from '@subwallet/extension-base/koni/api/contract-
 import { _BALANCE_CHAIN_GROUP, _MANTA_ZK_CHAIN_GROUP, _ZK_ASSET_PREFIX } from '@subwallet/extension-base/services/chain-service/constants';
 import { _EvmApi, _SubstrateAdapterSubscriptionArgs, _SubstrateApi } from '@subwallet/extension-base/services/chain-service/types';
 import { _checkSmartContractSupportByChain, _getAssetExistentialDeposit, _getChainExistentialDeposit, _getChainNativeTokenSlug, _getContractAddressOfToken, _getTokenOnChainAssetId, _getTokenOnChainInfo, _getTokenTypesSupportedByChain, _getXcmAssetMultilocation, _isBridgedToken, _isChainEvmCompatible } from '@subwallet/extension-base/services/chain-service/utils';
-import { getTaoToAlphaMapping, TaoStakeInfo } from '@subwallet/extension-base/services/earning-service/handlers/native-staking/tao';
+import { getTaoToAlphaMapping } from '@subwallet/extension-base/services/earning-service/handlers/native-staking/dtao';
+import { TaoStakeInfo } from '@subwallet/extension-base/services/earning-service/handlers/native-staking/tao';
 import { BalanceItem, SubscribeBasePalletBalance, SubscribeSubstratePalletBalance } from '@subwallet/extension-base/types';
 import { filterAssetsByChainAndType } from '@subwallet/extension-base/utils';
 import BigN from 'bignumber.js';
@@ -145,7 +146,7 @@ const subscribeWithSystemAccountPallet = async ({ addresses, callback, chainInfo
 
   let bittensorStakingBalances: BigN[] = new Array<BigN>(addresses.length).fill(new BigN(0));
 
-  if (['bittensor'].includes(chainInfo.slug)) {
+  if (['bittensor', 'bittensor_testnet'].includes(chainInfo.slug)) {
     bittensorStakingBalances = await Promise.all(addresses.map(async (address) => {
       const stakeInfo = (await substrateApi.api.call.stakeInfoRuntimeApi.getStakeInfoForColdkey(address)).toJSON() as Record<string, TaoStakeInfo> | undefined;
       const price = await getTaoToAlphaMapping(substrateApi);