+++ packages/extension-koni-ui/src/Popup/BuyTokens.tsx
@@ -3,16 +3,16 @@
 
 import { Resolver } from '@subwallet/extension-base/background/types';
 import { _getOriginChainOfAsset } from '@subwallet/extension-base/services/chain-service/utils';
-import { AccountProxy, BuyTokenInfo } from '@subwallet/extension-base/types';
+import { AccountProxy, BuyServiceInfo, BuyTokenInfo, SupportService } from '@subwallet/extension-base/types';
 import { detectTranslate, isAccountAll } from '@subwallet/extension-base/utils';
 import { AccountAddressSelector, baseServiceItems, Layout, PageWrapper, ServiceItem } from '@subwallet/extension-koni-ui/components';
 import { ServiceSelector } from '@subwallet/extension-koni-ui/components/Field/BuyTokens/ServiceSelector';
 import { TokenItemType, TokenSelector } from '@subwallet/extension-koni-ui/components/Field/TokenSelector';
 import { useAssetChecker, useDefaultNavigate, useGetChainSlugsByAccount, useNotification, useReformatAddress, useTranslation } from '@subwallet/extension-koni-ui/hooks';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
-import { AccountAddressItemType, BuyServiceInfo, CreateBuyOrderFunction, SupportService, ThemeProps } from '@subwallet/extension-koni-ui/types';
+import { AccountAddressItemType, CreateBuyOrderFunction, ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { BuyTokensParam } from '@subwallet/extension-koni-ui/types/navigation';
-import { createBanxaOrder, createCoinbaseOrder, createTransakOrder, noop, openInNewTab } from '@subwallet/extension-koni-ui/utils';
+import { createBanxaOrder, createCoinbaseOrder, createMeldOrder, createTransakOrder, noop, openInNewTab } from '@subwallet/extension-koni-ui/utils';
 import reformatAddress from '@subwallet/extension-koni-ui/utils/account/reformatAddress';
 import { Button, Form, Icon, ModalContext, SwModal, SwSubHeader } from '@subwallet/react-ui';
 import CN from 'classnames';
@@ -72,6 +72,7 @@ function Component ({ className, currentAccountProxy }: ComponentProps) {
   const { assetRegistry } = useSelector((state: RootState) => state.assetRegistry);
   const { walletReference } = useSelector((state: RootState) => state.settings);
   const { services, tokens } = useSelector((state: RootState) => state.buyService);
+
   const checkAsset = useAssetChecker();
   const allowedChains = useGetChainSlugsByAccount();
   const getReformatAddress = useReformatAddress();
@@ -101,7 +102,8 @@ function Component ({ className, currentAccountProxy }: ComponentProps) {
     banxa: false,
     onramper: false,
     moonpay: false,
-    coinbase: false
+    coinbase: false,
+    meld: false
   });
 
   const selectedAddress = Form.useWatch('address', form);
@@ -257,6 +259,9 @@ function Component ({ className, currentAccountProxy }: ComponentProps) {
         break;
       case 'coinbase':
         urlPromise = createCoinbaseOrder;
+        break;
+      case 'meld':
+        urlPromise = createMeldOrder;
         break;
     }