+++ packages/extension-koni-ui/src/Popup/Transaction/variants/Unbond.tsx
@@ -6,7 +6,7 @@ import { AmountData, ExtrinsicType, NominationInfo } from '@subwallet/extension-
 import { getValidatorLabel } from '@subwallet/extension-base/koni/api/staking/bonding/utils';
 import { _STAKING_CHAIN_GROUP } from '@subwallet/extension-base/services/earning-service/constants';
 import { isActionFromValidator } from '@subwallet/extension-base/services/earning-service/utils';
-import { AccountJson, RequestYieldLeave, SpecialYieldPoolMetadata, YieldPoolType, YieldPositionInfo } from '@subwallet/extension-base/types';
+import { AccountJson, RequestYieldLeave, SpecialYieldPoolMetadata, SubnetYieldPositionInfo, YieldPoolType, YieldPositionInfo } from '@subwallet/extension-base/types';
 import { AccountSelector, AlertBox, AmountInput, HiddenInput, InstructionItem, NominationSelector } from '@subwallet/extension-koni-ui/components';
 import { BN_ZERO, UNSTAKE_ALERT_DATA, UNSTAKE_BIFROST_ALERT_DATA, UNSTAKE_BITTENSOR_ALERT_DATA } from '@subwallet/extension-koni-ui/constants';
 import { MktCampaignModalContext } from '@subwallet/extension-koni-ui/contexts/MktCampaignModalContext';
@@ -94,7 +94,7 @@ const Component: React.FC = () => {
 
   const bondedAsset = useGetChainAssetInfo(bondedSlug || poolInfo.metadata.inputAsset);
   const decimals = bondedAsset?.decimals || 0;
-  const symbol = bondedAsset?.symbol || '';
+  const symbol = (positionInfo as SubnetYieldPositionInfo).subnetData?.subnetSymbol || bondedAsset?.symbol || '';
   const altAsset = useGetChainAssetInfo((poolInfo?.metadata as SpecialYieldPoolMetadata)?.altInputAssets);
   const altSymbol = altAsset?.symbol || '';
 
@@ -129,6 +129,10 @@ const Component: React.FC = () => {
         const exchaneRate = poolInfo.statistic?.assetEarning.find((item) => item.slug === input)?.exchangeRate || 1;
 
         return new BigN(positionInfo?.activeStake || '0').multipliedBy(exchaneRate).toFixed(0);
+      }
+
+      case YieldPoolType.SUBNET_STAKING: {
+        return selectedValidator?.activeStake || '0';
       }
 
       case YieldPoolType.LIQUID_STAKING:
@@ -338,7 +342,7 @@ const Component: React.FC = () => {
   }, [poolType, setCustomScreenTitle, t]);
 
   const exType = useMemo(() => {
-    if (poolType === YieldPoolType.NOMINATION_POOL || poolType === YieldPoolType.NATIVE_STAKING) {
+    if (poolType === YieldPoolType.NOMINATION_POOL || poolType === YieldPoolType.NATIVE_STAKING || poolType === YieldPoolType.SUBNET_STAKING) {
       return ExtrinsicType.STAKING_UNBOND;
     }
 
@@ -365,7 +369,7 @@ const Component: React.FC = () => {
 
   const unstakeAlertData = poolChain === 'bifrost_dot'
     ? UNSTAKE_BIFROST_ALERT_DATA
-    : poolChain === 'bittensor' ? UNSTAKE_BITTENSOR_ALERT_DATA : UNSTAKE_ALERT_DATA;
+    : poolChain.startsWith('bittensor') ? UNSTAKE_BITTENSOR_ALERT_DATA : UNSTAKE_ALERT_DATA;
 
   return (
     <>
@@ -410,6 +414,7 @@ const Component: React.FC = () => {
               label={t(`Select ${handleValidatorLabel}`)}
               networkPrefix={networkPrefix}
               nominators={nominators}
+              poolInfo={poolInfo}
             />
           </Form.Item>