+++ packages/extension-koni-ui/src/components/Earning/EarningPositionItem.tsx
@@ -1,7 +1,9 @@
 // Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { YieldPoolType } from '@subwallet/extension-base/types';
 import { BN_TEN } from '@subwallet/extension-base/utils';
+import DefaultLogosMap from '@subwallet/extension-koni-ui/assets/logo';
 import { NetworkTag } from '@subwallet/extension-koni-ui/components';
 import EarningTypeTag from '@subwallet/extension-koni-ui/components/Earning/EarningTypeTag';
 import { useSelector } from '@subwallet/extension-koni-ui/hooks';
@@ -51,23 +53,42 @@ const Component: React.FC<Props> = (props: Props) => {
     return chainInfoMap[positionInfo.chain].isTestnet;
   }, [chainInfoMap, positionInfo.chain]);
 
+  const subnetShortName = useMemo(() => {
+    return positionInfo.subnetData?.subnetShortName ? `(${positionInfo.subnetData.subnetShortName})` : '';
+  }, [positionInfo.subnetData?.subnetShortName]);
+
+  const isSubnetStaking = useMemo(() => [YieldPoolType.SUBNET_STAKING].includes(poolInfo.type) && !poolInfo.slug.includes('testnet'), [poolInfo.slug, poolInfo.type]);
+
   return (
     <div
       className={CN(className)}
       onClick={onClick}
     >
-      <div className={'__item-left-part'}>
-        <Logo
-          className={'__item-logo'}
-          isShowSubLogo={true}
-          size={40}
-          subNetwork={poolInfo.metadata.logo || poolInfo.chain}
-          token={balanceToken.toLowerCase()}
-        />
-
+      <div className='__item-left-part'>
+        {!isSubnetStaking || !DefaultLogosMap[`subnet-${poolInfo.metadata.subnetData?.netuid || 0}`]
+          ? (
+            <Logo
+              className='__item-logo'
+              isShowSubLogo={true}
+              size={40}
+              subNetwork={poolInfo.metadata.logo || poolInfo.chain}
+              token={balanceToken.toLowerCase()}
+            />
+          )
+          : (
+            <Logo
+              className='__item-logo'
+              isShowSubLogo={false}
+              network={`subnet-${poolInfo.metadata.subnetData?.netuid || 0}`}
+              shape='squircle'
+              size={40}
+            />
+          )}
         <div className='__item-lines-container'>
           <div className='__item-line-1'>
-            <div className='__item-name'>{poolName}</div>
+            <div className='__item-name'> { poolName }
+              <span className='__subnet-short-name'> {subnetShortName} </span>
+            </div>
 
             {
               !_isRelatedToAstar && (
@@ -188,6 +209,9 @@ const EarningPositionItem = styled(Component)<Props>(({ theme: { token } }: Prop
       fontWeight: token.headingFontWeight,
       overflow: 'hidden',
       textOverflow: 'ellipsis'
+    },
+    '.__subnet-short-name': {
+      color: token.colorTextLight4
     },
 
     '.__item-balance-value': {