+++ packages/subwallet-api-sdk/src/modules/swapApi.ts
@@ -39,6 +39,11 @@ export interface SwapRequest {
   currentQuote?: SwapProvider
 }
 
+export interface HydrationRateRequest {
+  address: string;
+  pair: SwapPair;
+}
+
 export type SwapRate = number;
 
 export interface QuoteAskResponse {
@@ -125,6 +130,35 @@ export class SwapApi {
       return response.result;
     } catch (error) {
       throw new Error(`Failed to fetch swap quote: ${(error as Error).message}`);
+    }
+  }
+
+  async getHydrationRate (hydrationRateRequest: HydrationRateRequest): Promise<number | undefined> {
+    const url = `${this.baseUrl}/swap/hydration-rate`;
+
+    try {
+      const rawResponse = await fetch(url, {
+        method: 'POST',
+        headers: {
+          accept: 'application/json',
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ hydrationRateRequest })
+      });
+
+      const response = await rawResponse.json() as SubWalletResponse<{ rate: number }>;
+
+      if (response.statusCode !== 200) {
+        console.error(response.message);
+
+        return undefined;
+      }
+
+      return response.result.rate;
+    } catch (error) {
+      console.error(`Failed to fetch swap quote: ${(error as Error).message}`);
+
+      return undefined;
     }
   }
 }